move around fixtures
